Fix test issues.
Fix formatting and correctness issues in the tests.

Error(f) is like Log + Fatal, so tests end after an Error(f). These
should be converted to Logf's in the future if we expect the additive
output to be useful. Error(f) should be reserved for when the test can
no longer proceed.